+++ data/normalized/CHANGELOG.md
@@ -1,13 +1,14 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-08-21T06:39:22.722Z
+Generated: 2026-08-22T06:33:34.363Z
 
 - 2026-08-01 • Microsoft Defender for Cloud • Deprecation • Date Category Update August 17, 2026 Upcoming deprecation Classic Defender for SQL
 - 2026-08-01 • Microsoft Defender for Endpoint • Preview • Preview/GA
 - 2026-08-01 • Microsoft Defender for Endpoint • Ikke oppgitt • GA Release version 20.126062.12.0 released: See enhancements and features for this release
 - 2026-08-01 • Microsoft Defender for Endpoint • Preview • Preview
 - 2026-08-01 • Microsoft Defender for Endpoint • Ikke oppgitt • GA Release version 20.126062.11.0 released: See enhancements and features for this release
-- 2026-08-01 • Microsoft Purview • Ikke oppgitt • Sensitivity labels
+- 2026-08-01 • Microsoft Purview • Ikke oppgitt • New : Before enforcing an auto-labeling policy, run it in simulation mode
+- 2026-08-01 • Microsoft Purview • Ikke oppgitt • New : The Insights tab in the policy details panel provides an
 - 2026-08-01 • Microsoft Sentinel • Preview • New data sources for UEBA behaviors and anomaly detection (Preview)
 - 2026-08-01 • Microsoft Sentinel • Preview • UEBA anomalies on behaviors (Preview)
 - 2026-08-01 • Microsoft Sentinel • Ikke oppgitt • UEBA anomaly detection now supports Check Point, Fortinet FortiGate, and Zscaler firewall, VPN, and web proxy events from the
@@ -21,4 +22,3 @@ Generated: 2026-08-21T06:39:22.722Z
 - 2026-07-01 • Microsoft Defender for Endpoint • Preview • Preview/GA
 - 2026-07-01 • Microsoft Defender for Endpoint • Ikke oppgitt • GA Release version 20.126062.9.0 released: See enhancements and features for this release
 - 2026-07-01 • Microsoft Defender for Endpoint • Ikke oppgitt • GA Release version 1.1.78290102 released: See enhancements and features for this release
-- 2026-07-01 • Microsoft Defender for Endpoint • Preview • Preview